+++ firmware/hdl/payload/frame.vhd
@@ -1,6 +1,7 @@
 -- Description:
 -- Contains the "framework" of GT-logic (all parts, except GTL and FDL).
 
+-- HB 2022-03-22: v1.3.0 - output ports bcres_d and bcres_d_FDL not used anymore (not used in mp7_payload.vhd). Signals bcres and bcres_outputmux not used anymore. Updated tcm.vhd (input port bcres_d_FDL not used anymore).
 -- HB 2021-06-16: v1.2.4 - implemented selectors (set in gtl_pkg.vhd) for "scouting" (in output_mux.vhd) and use of input data spymem.
 -- HB 2017-10-10: v1.2.3 - bug fix "simmem_in_use_i" input of spytrig.
 -- HB 2017-10-10: v1.2.2 - removed mux control register ("mux_ctrl_regs_1"), used fixed values for output mux inputs ("mux_ctrl").
@@ -62,8 +63,8 @@ entity frame is
         oc0 : in std_logic;
         start : in std_logic;
         l1a : in std_logic;
-        bcres_d : out std_logic;
-        bcres_d_FDL : out std_logic;
+--         bcres_d : out std_logic;
+--         bcres_d_FDL : out std_logic;
         start_lumisection : out std_logic;
         lane_data_in : in ldata(NR_LANES-1 downto 0);
         lane_data_out : out ldata(NR_LANES-1 downto 0);
@@ -105,10 +106,10 @@ architecture rtl of frame is
 
     signal lmp_lhc_data_o   : lhc_data_t; -- lhc_data output of lane mapping process
 
-    signal bcres : std_logic; -- NOT USED, "bc0" of mp7_ttc is used instead of "bcres"
-    signal bcres_d_int : std_logic; -- delayed version of bcres
-    signal bcres_d_FDL_int : std_logic; -- delayed version of bcres for FDL
-    signal bcres_outputmux : std_logic; -- non-delayed version of bcres for output mux
+--     signal bcres : std_logic; -- NOT USED, "bc0" of mp7_ttc is used instead of "bcres"
+    signal bc0_d_int : std_logic; -- delayed version of bcres
+--     signal bcres_d_FDL_int : std_logic; -- delayed version of bcres for FDL
+--     signal bcres_outputmux : std_logic; -- non-delayed version of bcres for output mux
 
     --TCM signals
     signal bx_nr : bx_nr_t;
@@ -221,8 +222,8 @@ architecture rtl of frame is
             oc0               => oc0_d_int,
             start             => start_d_int,
             l1a_sync          => l1a,
-            bcres_d           => bcres_d_int,
-            bcres_d_FDL       => bcres_d_FDL_int,
+            bcres_d           => bc0_d_int,
+--             bcres_d_FDL       => bcres_d_FDL_int,
             sw_reg_in         => rb2tcm,
             sw_reg_out        => tcm2rb,
             bx_nr             => bx_nr,
@@ -262,24 +263,26 @@ architecture rtl of frame is
     sync_proc_i : process (lhc_clk, lhc_rst)
     begin
         if lhc_rst = RST_ACT then
-            bcres_d_int <= '0';
+            bc0_d_int <= '0';
+--             bcres_d_int <= '0';
             ec0_d_int <= '0';
             oc0_d_int <= '0';
             start_d_int <= '0';
-            bcres_d_FDL_int <= '0';
-            bcres_outputmux <= '0';
+--             bcres_d_FDL_int <= '0';
+--             bcres_outputmux <= '0';
         elsif rising_edge(lhc_clk) then
-            bcres_d_int  <= bc0;
+            bc0_d_int  <= bc0;
+--             bcres_d_int  <= bc0;
             ec0_d_int <= ec0;
             oc0_d_int <= oc0;
             start_d_int <= start;
-            bcres_d_FDL_int <= bc0;
-            bcres_outputmux <= bc0;
+--             bcres_d_FDL_int <= bc0;
+--             bcres_outputmux <= bc0;
         end if;
     end process;
 
-    bcres_d_FDL <= bcres_d_FDL_int;
-    bcres_d <= bcres_d_int;
+--     bcres_d_FDL <= bcres_d_FDL_int;
+--     bcres_d <= bcres_d_int;
 
 --===============================================================================================--
 --                           SIM/SPY MEMORY